fix mobile tablet review feedback
